Enable "Local" provider type (#274)
This change is a step towards creating a local, AWS-like provider to enable more flexibility for both development and testing. AWS-like means the local directory structure is setup exactly like a cost usage report in an S3 bucket and the reports that it can process must conform to the AWS structure.