Removing expired report data on regular monthly schedule (#122)
This is a continuation of PR #119. Now This same logic kicked off by the DEL /api/v1/expired_data call will be ran in a beat schedule that runs on the first of the month. At midnight the expired report data will be cleared.

A new environment variable REMOVE_EXPIRED_REPORT_DATA_ON_DAY was added so we can adjust what day of the month report data cleaning will take place. If it is set to 0 then report cleaning will be disabled.