Load earlier months of data when adding new providers. (#204)
It is desired that multiple months of data is initially pulled into masu/koku when a new provider account is added. This change will now download and process the current and previous month by default when a new account is added. New environment variables are added so we can adjust these in production (INITIAL_INGEST_OVERRIDE and INITIAL_INGEST_NUM_MONTHS).