Updated OpenAPI for current server implementation
Generating the swagger client against the current specification doesn't
actually work. The previous and next fields are refined as required,
yet if there is only one record, deserialization fails as previous
and next are not allowed to be None.

The results attribute was also not present in the specification and so
the deserialization omitted it when using the python client.

* Removed the required marker for previous and next in the ListPagination
  object definition.
* Added results with type object to the ListPagination object.