Break out summary tasks (#445)
This splits daily table summarization from the final summary table summarization. Both are still accomplished in the report summary updaters, but now in separate methods. This is precursor work to calculating charge/derived cost on the daily tables. Note that derived cost calculations and OCP Cloud are untouched here, and are being saved for follow-up work.