Revert back to using temp tables for processing GCP reports.
It turns out that using the
	self._delete_line_items()
	self._save_to_db()
pattern for processing GCP reports will not work.

_delete_line_items will remove all line items associated with a manifest.
Since the GCP manifests are monthly, but the report files are daily,
this would remove all line items processed for previous reports.

the insert conflict columns pattern works well for GCP, so I am reverting
back to that.

Add logic to deal with duplicate line item rows