Add more decimal precision to calculated decimal fields.
This is needed because the cost in reporting_ocpawscostlineitem_project_daily_summary (among others) is
calculated from daily.unblended_cost/daily.shared_pods. When this is stored in the
temporary table reporting_ocpawscostlineitem_project_daily_summary_{{uuid}}
the field type in postgres is numeric, which does not coerce input values to any precision[0].
But when this is copied over to reporting_ocpawscostlineitem_project_daily_summary
the type is coerced to numeric(24,9). Which results in some loss of precision.

Since Django does not allow us to specify a unlimited Decimal Field, an alternative solution is to
just store the calculated decimals at a higher precision than non-calculated decimals.
This way we should be able to minimize rounding errors.

[0]https://www.postgresql.org/docs/9.3/datatype-numeric.html#DATATYPE-NUMERIC-DECIMAL