Fix SRE CI by adding GIT_COMMIT
The SRE CI job fails due to using previous git commit when updating
a PR. This causes problems because the git commit is used to
get labels and according to them to run tests and build image.

To fix this I added GIT_COMMIT variable that is used in the
CI script to fetch labels from Github.

I set the GIT_COMMIT to the current HEAD of a PR.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>