Change bind mount directory creation
Trino runs as root inside the container, so it is not possible to
recursively change owenership of the trino data files. Leave that
alone for now.

Only set ownership and mode of the directories that are bind mounts
and not their contents.

Remove a minio build in the make target setting permissions. That
seems to have been added to that target by mistake.