Only read currency API data from a good response
Check the response before decoding the response data. If the response is
unsuccessful, there will be no data, which results in an different exception
from the .json() method. This obscures the real problem, which is that the API
did not return data.

Add retry with exponential backoff to add resiliency.