Warn when a partition already exists for OCP processing
This appears to be a problem only with OCP processing and only when two workers
are running the same workload practiaclly at the same exact time. This may
indicate a problem with our worker caching not preventing simultaneous
execution, or a problem with celery allowing duplicate tasks to run. We don’t
have enough information to determine that currently.

Since this is isolated to OCP processing, I fixed it only in the OPC processing
instead of making changes to update_summary_tables, which would affect all
provider types.