Update clone_schema SQL function when necessary
We may not end up wanting this to be so automatic, but this is a good way to
handle it if needed.

The only problem is getting the hash is a bit clunky since the contents of
clone_schema.sql do not exactly match the function definition once it is stored
in the database. I added the hash output to logs in order to make updating the
hash easier.