[COST-4515] Upgrade to Django 4.1 (#4853)
* upgrade to Django 4.1

* Update deprecated ‘warn’ method

* Remove deprecated test case method

* Remove deprecated Series.bool method

* Correct docstring and class attribute name

* Cleanup file handle

* Fix flaky test

This assert statement will always fail since the template schema does not yet
exists. This test will only pass if another test runs before this test creates
the template schema.

* Update clone_schema.sql

Do not define a custom sequence object since that is no longer required with the
move to identity columns.

Co-authored-by: Michael Skarbek <mskarbek@redhat.com>
Co-authored-by: David Nakabaale <dnakabaa@redhat.com>

* Update clone_schema SQL function when necessary

We may not end up wanting this to be so automatic, but this is a good way to
handle it if needed.

The only problem is getting the hash is a bit clunky since the contents of
clone_schema.sql do not exactly match the function definition once it is stored
in the database. I added the hash output to logs in order to make updating the
hash easier.

* Avoid unnecessary globals

* update comments in clone_schema func

* Update clone_schema hash

* add management cmd to migrate serial cols

* Update clone_schema.sql

Do not define a custom sequence object since that is no longer required with the
move to identity columns.

Co-authored-by: Michael Skarbek <mskarbek@redhat.com>
Co-authored-by: David Nakabaale <dnakabaa@redhat.com>

* Update license information

Add correct copyright and license information to management command.

Create project level licenses to reflect the use of multiple licenses within
the project.

* Improvements to serial to identity management command

- Use SQL composition rather than format strings to protect against SQL injection
  attacks. The risk is quite low, but it’s best to just do the right thing.
- Reorganize the code to avoid a nested function definiton and an unnecessary
  global variable that only needed to be local.

* Do not update clone_schema function automatically

* Add migration to update clone_schema function

* fixup! Improvements to serial to identity management command

---------

Co-authored-by: Sam Doran <sdoran@redhat.com>
Co-authored-by: Michael Skarbek <mskarbek@redhat.com>