Update redis cache settings
Ignore ConnectionError and retry instead of failing immediately. This was the
configuration we had before with django_redis and the IGNORE_EXCEPTIONS setting.

With the move to Django’s RedisCache, it just passes through the OPTIONS
directly to the backend library, which in this case is redis-py.

redis-py actually provides finer grained controls around handling and retries.
It’s possible to select which errors are ignored and which retry strategy
is used.

I chose to ignore only ConnectionError and retry with exponential backoff.
We are only seeing ConnectionError currently. This can be changed as needed.

I also cleaned up other redis settings so that values are only defined in
one place and made the long integers easier to read.