Rename fields storing data and add a unit field
We decided to store the data in GB (base ten gigabytes) to align with other
cost types. Also add a field to store the unit value so that we can return
this in the API.

We may need to add a column to OCPUsageLineItemDailySummary for this unit, but
that may get messy since there are multiple values in the row. That model is
just a hot mess. Not adding to it would be ideal.