Report correct PostgreSQL version
The version string digits changed after PostgreSQL 9. Update the version string
parsing to work with both version number schemas.

Change get_pg_engine_version method to return a string instead of list of ints.
This makes it easier for the caller.

This will need to be updated when PostgreSQL goes above version 99.

- Use a simpler SQL command to get the version string from the database
- Remove unecessary use of a global
- Add more test cases