Update image used by smoke test
Use the Konflux release image. Since this test is running against
main, it doesn't need to wait for an image build since the
image should already be built.

By default, the image tag is the 7 character SHA of HEAD, which
lines up with how the Konflux release images are tagged.